@@ -1,10 +1,10 @@
-package com.example.ApiGateWay;
+package com.example.demo;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApiGateWayApplicationTests {
+class DemoApplicationTests {
 
 	@Test
 	void contextLoads() {
