@@ -1,18 +1,15 @@
-package com.example.ApiGateWay;
+package com.example.MicroServer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
 
 @SpringBootApplication
 @EnableEurekaClient
-
-public class ApiGateWayApplication {
+public class MicroServerApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ApiGateWayApplication.class, args);
+		SpringApplication.run(MicroServerApplication.class, args);
 	}
 
 }
